Adds a dealer module for Dealer class
